pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["Matheus Henrique <matheus2hcosta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^1.2.0"
pylint = "^2.17.2"
pika = "^1.3.1"
types-pika = "^1.2.0b1"
fastapi = "^0.95.1"
pymongo = "^4.3.3"
akeyless = "^3.3.0"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
names = "^0.3.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[mypy]
ignore_missing_imports = true

